request types controller and some dir changes

+++ eduflow/src/main/java/com/final_project/eduflow/DataAccess/RequestTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.final_project.eduflow.Data.Entities.RequestType;
 
-public interface RequestTypeRepository extends JpaRepository<RequestType, Integer>{
+public interface RequestTypeRepository extends JpaRepository<RequestType, Long>{
     
 }
